requirements.txt
Flask==2.3.2
openai==1.3.6
python-dotenv==1.0.0
requests==2.31.0
sentence-transformers==2.2.2
PyPDF2==3.0.1
